Fix dcraw warnings in linear DNG file. Patch from Dave Coffin that will appear in next version of dcraw.
Do not use ppg_inderpolation if colors==4 (as dcraw does).

Fix the sometimes missing green channel in the raw histogram.

Show only relevant interpolations in the combo-box.

+++ dcraw.cc
@@ -867,7 +867,7 @@ void CLASS ljpeg_row (int jrow, struct jhead *jh)
     for (c=0; c < jh->clrs; c++) {
       diff = ljpeg_diff (jh->huff[c]);
       *outp = diff = col ? outp[-jh->clrs]+diff : (jh->vpred[c] += diff);
-      if (diff >> jh->bits) derror();
+      if ((ushort) diff >> jh->bits) derror();
       outp++;
     }
 }
